Extract repro for #10803
